learn about policy and add update question feature

+++ app/Http/Controllers/QuestionController.php
@@ -4,73 +4,27 @@ namespace App\Http\Controllers;
 
 use App\Models\Question;
 use App\Services\QuestionService;
+use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 
 class QuestionController extends Controller
 {
-    public function index(){
-        $questions = Question::withCount("answers")
-                        ->filter(request(['search','tag']))
-                        ->latest()
-                        ->paginate(10)
-                        ->through(function($q){
-                            $q->authorized = $q->user_id == auth()->id();
-                            return $q;
-                        })
-                        ->withQueryString();
+    public function index(QuestionService $questionService){
         return inertia('Questions/Index', [
             'filters' => request(['search','tag']),
-            'questions' => Inertia::scroll($questions),
+            'questions' => Inertia::scroll($questionService->getQuestions()),
         ]);
     }
-    public function show(Question $question){
-        $sort = request('sort');
-        $answers = $question->answers()
-                    ->when($sort === 'latest', function($query){$query->latest();})
-                    ->paginate(3)
-                    ->withQueryString();
-                    
+    public function show(Question $question, QuestionService $questionService){
         return inertia('Questions/Detail',[
             'question' => $question,
-            'answers' => $answers,
-            'sort' => $sort,
-            'relatedQuestions' => $this->getRelatedQuestions($question),
+            'answers' => $questionService->getAnswers($question),
+            'sort' => request('sort'),
+            'relatedQuestions' => $questionService->getRelatedQuestions($question),
         ]);
     }
-    public function getRelatedQuestions($question){
-        return cache()->remember("relatedQuestions_".$question->id,now()->addMinute(),function() use($question){
-            // if no tags, get latest questions
-            if ($question->tags->isEmpty()) {
-                return $this->getLatestQuestions($question);
-            }
-            // if has tags, get questions with same tags
-            $tags=$question->tags->pluck('name');
-            $result = Question::whereHas('tags',function($query) use($tags){
-                $query->whereIn('name',$tags);
-            })->where('id','!=',$question->id)->inRandomOrder()->limit(3)->get();
-            // if no result, get latest questions
-            if($result->isEmpty()){
-                $result = $this->getLatestQuestions($question);
-            }elseif ($result->count() <3) {
-                $needed = 3 - $result->count();
-                Question::where('id','!=',$question->id)
-                        ->whereNotIn('id',$result->pluck('id')) //exclude already related questions
-                        ->latest()
-                        ->limit($needed)
-                        ->get()
-                        ->merge($result); //merge with existing results
-            }
-            return $result;
-        });
-    }
-    public function getLatestQuestions($question){
-        return Question::where('id','!=',$question->id)
-                ->latest()
-                ->limit(3)
-                ->get();
-    }
     public function create(){
-        return inertia('Questions/Create');
+        return inertia('Questions/QuestionForm');
     }
     public function store(QuestionService $questionService){
         $newQuestion=request()->validate([
@@ -81,7 +35,23 @@ class QuestionController extends Controller
         return redirect()->route('questions.detail', $newQuestion->id);
     }
     public function destroy(Question $question){
+        Gate::authorize('delete',$question);
         $question->delete();
         return redirect()->route('home');
+    }
+    public function edit(Question $question){
+        Gate::authorize('update',$question);
+        return inertia('Questions/QuestionForm',[
+            'question' => $question,
+        ]);
+    }
+    public function update(Question $question){
+        Gate::authorize('update',$question);
+        $updateQuestion=request()->validate([
+            'title' => "required| min: 5",
+            'body' => "required| min: 20"
+        ]);
+        $question->update($updateQuestion);
+        return redirect()->route('questions.detail', $question->id);
     }
 }
